+++ src/main/java/pl/com/mike/developer/domain/developer/PremiseData.java
@@ -17,7 +17,6 @@ public class PremiseData implements Serializable {
     @Column(name = "surface_sq_m")
     private Double surfacePerSqMeter;
     @Column(name = "price_of_sq_m")
-
     private Double pricePerSqMeter;
     @Column(name = "price_total")
     private BigDecimal totalPrice;
@@ -37,12 +36,12 @@ public class PremiseData implements Serializable {
     @Column(name = "is_loggia")
     private Boolean isLoggia;
     @Column(name = "building_id")
-    private Integer buildingId;
+    private Long buildingId;
 
     public PremiseData() {
     }
 
-    public PremiseData(Long id, String type, Integer number, Integer floor, Double surfacePerSqMeter, Double pricePerSqMeter, BigDecimal totalPrice, Integer numberOfRooms, String technicalStatus, String salesStatus, String exposure, Boolean isBalcony, Boolean isGarden, Boolean isTerrace, Boolean isLoggia, Integer buildingId) {
+    public PremiseData(Long id, String type, Integer number, Integer floor, Double surfacePerSqMeter, Double pricePerSqMeter, BigDecimal totalPrice, Integer numberOfRooms, String technicalStatus, String salesStatus, String exposure, Boolean isBalcony, Boolean isGarden, Boolean isTerrace, Boolean isLoggia, Long buildingId) {
         this.id = id;
         this.type = type;
         this.number = number;
@@ -181,11 +180,12 @@ public class PremiseData implements Serializable {
         isLoggia = loggia;
     }
 
-    public Integer getBuildingId() {
+
+    public Long getBuildingId() {
         return buildingId;
     }
 
-    public void setBuildingId(Integer buildingId) {
+    public void setBuildingId(Long buildingId) {
         this.buildingId = buildingId;
     }
